move to standalone module
